requirements.txt
asgiref==3.8.1
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==43.0.0
Django==5.0.7
django-oauth-toolkit==2.4.0
django-otp==1.5.1
djangorestframework==3.15.2
idna==3.7
jwcrypto==1.5.6
oauthlib==3.2.2
pillow==10.4.0
pycparser==2.22
pypng==0.20220715.0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
qrcode==7.4.2
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
